set up some tests

+++ pragmatic-programmer/json-yaml.rb
@@ -4,14 +4,26 @@ require 'yaml'
 directory = ARGV[0] ? ARGV[0] : Dir.pwd
 delete = ARGV[1] == "true" ? true : false
 
-yaml_files = Dir.glob("#{directory}/*{yaml,yml}")
-yaml_files.each do |file|
-  
-  json_text = json.pretty_generate(yaml.load_file(file))
-  filename = "#{file.split('.').first}.json"
-  File.open(filename, 'w').write(json_text)
+def jsonify_yaml_file(filename)
+  JSON.pretty_generate(YAML.load_file(filename))
+end
 
-  if delete
-    File.delete(file)
+def convert_filename(filename)
+  splitname = filename.split('.')
+  range = (0...splitname.length - 1)
+  "#{splitname[range].join('.')}.json"
+end
+
+if __FILE__ == $0
+  yaml_files = Dir.glob("#{directory}/*{yaml,yml}")
+  yaml_files.each do |file|
+
+    json_text = jsonify_yaml_file(file)
+    new_name = convert_filename(filename)
+    File.write(new_name, jsonified_text)
+
+    if delete
+      File.delete(file)
+    end
   end
 end
